move file to vendor directory
